Style: 검색 컴포넌트 분리
