+++ src/components/Header/HeaderParts/index.ts
@@ -1,5 +1,5 @@
 export { default as CartButton } from './CartButton';
 export { default as LoginButton } from './LoginButton';
 export { default as LogoutButton } from './LogoutButton';
-export { default as SearchBox } from './SearchBox';
+export { default as SearchButton } from './SearchButton';
 export { default as ShopButton } from './ShopButton';
