+++ src/components/SearchBox/SearchHistory.module.scss
