+++ src/components/SearchBox/SearchHistory.tsx
