+++ src/components/SearchBox/SearchSuggestion.module.scss
